requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    ./bin/compile_dependencies.sh
#
-e git+git://github.com/maykinmedia/mail-editor@34ee3c9709ac16cf5bf6dee4ea3619c22419b886#egg=mail_editor
    # via -r requirements/base.in
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django-axes==5.20.0
    # via -r requirements/base.in
django-choices==1.7.2
    # via
    #   -r requirements/base.in
    #   mail-editor
    #   vng-api-common
    #   zgw-consumers
django-ckeditor==6.1.0
    # via mail-editor
django-cors-middleware==1.5.0
    # via -r requirements/base.in
django-filter==2.4.0
    # via vng-api-common
django-ipware==3.0.2
    # via django-axes
django-js-asset==1.2.2
    # via django-ckeditor
django-markup==1.5
    # via -r requirements/base.in
django-relativedelta==1.1.2
    # via zgw-consumers
django-rest-framework-condition==0.1.1
    # via vng-api-common
django-solo==1.1.5
    # via
    #   -r requirements/base.in
    #   vng-api-common
    #   zgw-consumers
django==2.2.24
    # via
    #   -r requirements/base.in
    #   django-axes
    #   django-choices
    #   django-filter
    #   django-markup
    #   django-relativedelta
    #   django-rest-framework-condition
    #   djangorestframework
    #   drf-nested-routers
    #   drf-yasg
    #   mail-editor
    #   vng-api-common
    #   zgw-consumers
djangorestframework-camel-case==1.2.0
    # via vng-api-common
djangorestframework==3.12.4
    # via
    #   -r requirements/base.in
    #   drf-nested-routers
    #   drf-yasg
    #   vng-api-common
drf-nested-routers==0.93.3
    # via vng-api-common
drf-yasg==1.20.0
    # via vng-api-common
faker==8.10.3
    # via zgw-consumers
gemma-zds-client==1.0.0
    # via
    #   vng-api-common
    #   zgw-consumers
idna==2.10
    # via requests
importlib-metadata==4.6.1
    # via markdown
inflection==0.5.1
    # via drf-yasg
iso-639==0.4.5
    # via vng-api-common
isodate==0.6.0
    # via vng-api-common
itypes==1.2.0
    # via coreapi
jinja2==3.0.1
    # via coreschema
markdown==3.3.4
    # via -r requirements/base.in
markupsafe==2.0.1
    # via jinja2
oyaml==1.0
    # via vng-api-common
packaging==21.0
    # via drf-yasg
psycopg2==2.8.6
    # via -r requirements/base.in
pyjwt==2.1.0
    # via
    #   gemma-zds-client
    #   vng-api-common
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.1
    # via
    #   -r requirements/base.in
    #   django-relativedelta
    #   faker
python-dotenv==0.18.0
    # via -r requirements/base.in
pytz==2021.1
    # via
    #   -r requirements/base.in
    #   django
pyyaml==5.4.1
    # via
    #   gemma-zds-client
    #   oyaml
    #   vng-api-common
raven==6.10.0
    # via -r requirements/base.in
requests-mock==1.9.3
    # via zgw-consumers
requests==2.25.1
    # via
    #   coreapi
    #   gemma-zds-client
    #   requests-mock
    #   vng-api-common
    #   zgw-consumers
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
ruamel.yaml==0.17.10
    # via drf-yasg
six==1.16.0
    # via
    #   django-choices
    #   isodate
    #   python-dateutil
    #   requests-mock
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via faker
typing-extensions==3.10.0.0
    # via importlib-metadata
unidecode==1.2.0
    # via vng-api-common
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.6
    # via requests
uwsgi==2.0.19.1
    # via -r requirements/base.in
vng-api-common==1.6.2
    # via -r requirements/base.in
zgw-consumers==0.15.2
    # via -r requirements/base.in
zipp==3.5.0
    # via importlib-metadata
